backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
APScheduler
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
tqdm
pytz

